# api-contracts/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /thresholds/default:
    put:
      operationId: ThresholdsController_setDefaultThreshold
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetThresholdDefaultTypeRequestDto'
      responses:
        '200':
          description: ''
      tags:
        - Thresholds
  /thresholds/sensor/{sensorId}:
    put:
      operationId: ThresholdsController_setSensorThreshold
      parameters:
        - name: sensorId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetThresholdSensorRequestDto'
      responses:
        '200':
          description: ''
      tags:
        - Thresholds
    delete:
      operationId: ThresholdsController_deleteSensorThreshold
      parameters:
        - name: sensorId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Thresholds
  /thresholds/type/{sensorType}:
    delete:
      operationId: ThresholdsController_deleteThresholdType
      parameters:
        - name: sensorType
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Thresholds
  /users/{id}:
    get:
      operationId: UsersController_getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      summary: Get user details by id
      tags:
        - Users
    patch:
      operationId: UsersController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserResponseDto'
      summary: Update user in tenant
      tags:
        - Users
  /users:
    get:
      operationId: UsersController_getUsers
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
      summary: Get users for tenant
      tags:
        - Users
    post:
      operationId: UsersController_createUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponseDto'
      summary: Create user in tenant
      tags:
        - Users
  /users/bulk-delete:
    post:
      operationId: UsersController_deleteUsers
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteUserRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserResponseDto'
      summary: Delete users by ids
      tags:
        - Users
info:
  title: NoTIP Management API
  description: NoTIP Management API OpenAPI specification
  version: 1.2.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    SetThresholdDefaultTypeRequestDto:
      type: object
      properties:
        sensor_type:
          type: string
        min_value:
          type: number
        max_value:
          type: number
      required:
        - sensor_type
        - min_value
        - max_value
    SetThresholdSensorRequestDto:
      type: object
      properties:
        min_value:
          type: number
        max_value:
          type: number
        sensor_type:
          type: string
      required:
        - min_value
        - max_value
        - sensor_type
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - system_admin
            - tenant_admin
            - tenant_user
        last_access:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - email
        - role
        - last_access
    CreateUserRequestDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - system_admin
            - tenant_admin
            - tenant_user
        password:
          type: string
      required:
        - name
        - email
        - role
        - password
    CreateUserResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - system_admin
            - tenant_admin
            - tenant_user
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - created_at
    UpdateUserRequestDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - system_admin
            - tenant_admin
            - tenant_user
        permissions:
          type: array
          items:
            type: string
      required:
        - name
        - email
        - role
        - permissions
    UpdateUserResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - system_admin
            - tenant_admin
            - tenant_user
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - updated_at
    DeleteUserRequestDto:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
      required:
        - ids
    DeleteUserResponseDto:
      type: object
      properties:
        deleted:
          type: number
        failed:
          type: array
          items:
            type: string
      required:
        - deleted
        - failed